# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Gestione Prenotazioni Concessionaria
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Vettura {
  id: number;
  marca: string;
  modello: string;
  anno: number;
  targa: string;
  /** benzina | diesel | elettrica | ibrida | gpl */
  carburante: string;
  /** nuova | usata | km0 */
  stato: string;
  /** @nullable */
  colore?: string | null;
  /** @nullable */
  prezzo?: number | null;
  /** @nullable */
  km?: number | null;
  disponibile: boolean;
  /** @nullable */
  note?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateVetturaBody {
  marca: string;
  modello: string;
  anno: number;
  targa: string;
  carburante: string;
  stato: string;
  /** @nullable */
  colore?: string | null;
  /** @nullable */
  prezzo?: number | null;
  /** @nullable */
  km?: number | null;
  disponibile: boolean;
  /** @nullable */
  note?: string | null;
}

export interface UpdateVetturaBody {
  /** @nullable */
  marca?: string | null;
  /** @nullable */
  modello?: string | null;
  /** @nullable */
  anno?: number | null;
  /** @nullable */
  targa?: string | null;
  /** @nullable */
  carburante?: string | null;
  /** @nullable */
  stato?: string | null;
  /** @nullable */
  colore?: string | null;
  /** @nullable */
  prezzo?: number | null;
  /** @nullable */
  km?: number | null;
  /** @nullable */
  disponibile?: boolean | null;
  /** @nullable */
  note?: string | null;
}

export interface Cliente {
  id: number;
  nome: string;
  cognome: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  telefono?: string | null;
  /** @nullable */
  codiceFiscale?: string | null;
  /** @nullable */
  indirizzo?: string | null;
  /** @nullable */
  note?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface Prenotazione {
  id: number;
  vetturaId: number;
  clienteId: number;
  dataInizio: string;
  dataFine: string;
  /** attiva | completata | annullata | in_corso */
  stato: string;
  /** @nullable */
  note?: string | null;
  vettura: Vettura;
  cliente: Cliente;
  createdAt: string;
  updatedAt: string;
}

export interface Contratto {
  id: number;
  vetturaId: number;
  clienteId: number;
  /** @nullable */
  prenotazioneId?: number | null;
  numero: string;
  /** vendita | noleggio | leasing | permuta */
  tipo: string;
  dataContratto: string;
  /** @nullable */
  importo?: number | null;
  archiviato: boolean;
  /** @nullable */
  note?: string | null;
  vettura: Vettura;
  cliente: Cliente;
  createdAt: string;
  updatedAt: string;
}

export interface VetturaStorico {
  vettura: Vettura;
  prenotazioni: Prenotazione[];
  contratti: Contratto[];
}

export interface CreateClienteBody {
  nome: string;
  cognome: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  telefono?: string | null;
  /** @nullable */
  codiceFiscale?: string | null;
  /** @nullable */
  indirizzo?: string | null;
  /** @nullable */
  note?: string | null;
}

export interface UpdateClienteBody {
  /** @nullable */
  nome?: string | null;
  /** @nullable */
  cognome?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  telefono?: string | null;
  /** @nullable */
  codiceFiscale?: string | null;
  /** @nullable */
  indirizzo?: string | null;
  /** @nullable */
  note?: string | null;
}

export interface ClienteStorico {
  cliente: Cliente;
  prenotazioni: Prenotazione[];
  contratti: Contratto[];
}

export interface CreatePrenotazioneBody {
  vetturaId: number;
  clienteId: number;
  dataInizio: string;
  dataFine: string;
  stato: string;
  /** @nullable */
  note?: string | null;
}

export interface UpdatePrenotazioneBody {
  /** @nullable */
  vetturaId?: number | null;
  /** @nullable */
  clienteId?: number | null;
  /** @nullable */
  dataInizio?: string | null;
  /** @nullable */
  dataFine?: string | null;
  /** @nullable */
  stato?: string | null;
  /** @nullable */
  note?: string | null;
}

export interface CreateContrattoBody {
  vetturaId: number;
  clienteId: number;
  /** @nullable */
  prenotazioneId?: number | null;
  numero: string;
  tipo: string;
  dataContratto: string;
  /** @nullable */
  importo?: number | null;
  archiviato: boolean;
  /** @nullable */
  note?: string | null;
}

export interface UpdateContrattoBody {
  /** @nullable */
  numero?: string | null;
  /** @nullable */
  tipo?: string | null;
  /** @nullable */
  dataContratto?: string | null;
  /** @nullable */
  importo?: number | null;
  /** @nullable */
  archiviato?: boolean | null;
  /** @nullable */
  note?: string | null;
}

export interface CarburanteCount {
  carburante: string;
  count: number;
}

export interface StatoCount {
  stato: string;
  count: number;
}

export interface DashboardStats {
  totaleVetture: number;
  vettureDisponibili: number;
  vetturePrenotate: number;
  totaleClienti: number;
  prenotazioniAttive: number;
  prenotazioniMese: number;
  contrattiTotali: number;
  contrattiArchiviati: number;
  ripartizioneCarburante: CarburanteCount[];
  ripartizioneStato: StatoCount[];
}

export interface PrenotazioneCalendario {
  id: number;
  vetturaId: number;
  clienteId: number;
  dataInizio: string;
  dataFine: string;
  stato: string;
  vetturaNome: string;
  clienteNome: string;
  targa: string;
}

export type ListVettureParams = {
  /**
   * @nullable
   */
  carburante?: string | null;
  /**
   * @nullable
   */
  stato?: string | null;
  /**
   * @nullable
   */
  marca?: string | null;
  /**
   * @nullable
   */
  disponibile?: string | null;
};

export type ListClientiParams = {
  /**
   * @nullable
   */
  search?: string | null;
};

export type ListPrenotazioniParams = {
  /**
   * @nullable
   */
  vetturaId?: number | null;
  /**
   * @nullable
   */
  clienteId?: number | null;
  /**
   * @nullable
   */
  stato?: string | null;
  /**
   * @nullable
   */
  dal?: string | null;
  /**
   * @nullable
   */
  al?: string | null;
};

export type ListContrattiParams = {
  /**
   * @nullable
   */
  vetturaId?: number | null;
  /**
   * @nullable
   */
  clienteId?: number | null;
  /**
   * @nullable
   */
  archiviato?: string | null;
};

export type GetPrenotazioniCalendarioParams = {
  /**
   * @nullable
   */
  anno?: number | null;
  /**
   * @nullable
   */
  mese?: number | null;
};