# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Gestione Prenotazioni Concessionaria
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clienti:
    get:
      operationId: listClienti
      tags: [clienti]
      summary: Lista clienti
      parameters:
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Lista clienti
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cliente"
    post:
      operationId: createCliente
      tags: [clienti]
      summary: Crea nuovo cliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClienteBody"
      responses:
        "201":
          description: Cliente creato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
  /clienti/{id}:
    get:
      operationId: getCliente
      tags: [clienti]
      summary: Dettaglio cliente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cliente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
        "404":
          description: Not found
    patch:
      operationId: updateCliente
      tags: [clienti]
      summary: Aggiorna cliente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClienteBody"
      responses:
        "200":
          description: Cliente aggiornato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
        "404":
          description: Not found
    delete:
      operationId: deleteCliente
      tags: [clienti]
      summary: Elimina cliente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Eliminato
  /clienti/{id}/storico:
    get:
      operationId: getClienteStorico
      tags: [clienti]
      summary: Storico prenotazioni e contratti di un cliente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Storico cliente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClienteStorico"
...
  /contratti:
    get:
      operationId: listContratti
      tags: [contratti]
      summary: Lista contratti
      parameters:
        - name: vetturaId
          in: query
          schema:
            type: ["integer", "null"]
        - name: clienteId
          in: query
          schema:
            type: ["integer", "null"]
        - name: archiviato
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Lista contratti
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contratto"
    post:
      operationId: createContratto
      tags: [contratti]
      summary: Crea nuovo contratto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContrattoBody"
      responses:
        "201":
          description: Contratto creato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contratto"
  /contratti/{id}:
    get:
      operationId: getContratto
      tags: [contratti]
      summary: Dettaglio contratto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contratto
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contratto"
        "404":
          description: Not found
    patch:
      operationId: updateContratto
      tags: [contratti]
      summary: Aggiorna/archivia contratto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContrattoBody"
      responses:
        "200":
          description: Contratto aggiornato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contratto"
        "404":
          description: Not found
    delete:
      operationId: deleteContratto
      tags: [contratti]
      summary: Elimina contratto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Eliminato
...
components:
  schemas:
...
    Vettura:
      type: object
      properties:
        id:
          type: integer
        marca:
          type: string
        modello:
          type: string
        anno:
          type: integer
        targa:
          type: string
        carburante:
          type: string
          description: "benzina | diesel | elettrica | ibrida | gpl"
        stato:
          type: string
          description: "nuova | usata | km0"
        colore:
          type: ["string", "null"]
        prezzo:
          type: ["number", "null"]
        km:
          type: ["integer", "null"]
        disponibile:
          type: boolean
        note:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - marca
        - modello
        - anno
        - targa
        - carburante
        - stato
        - disponibile
        - createdAt
        - updatedAt
...
    Cliente:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        cognome:
          type: string
        email:
          type: ["string", "null"]
        telefono:
          type: ["string", "null"]
        codiceFiscale:
          type: ["string", "null"]
        indirizzo:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - nome
        - cognome
        - createdAt
        - updatedAt
    CreateClienteBody:
      type: object
      properties:
        nome:
          type: string
        cognome:
          type: string
        email:
          type: ["string", "null"]
        telefono:
          type: ["string", "null"]
        codiceFiscale:
          type: ["string", "null"]
        indirizzo:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
      required:
        - nome
        - cognome
    UpdateClienteBody:
      type: object
      properties:
        nome:
          type: ["string", "null"]
        cognome:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        telefono:
          type: ["string", "null"]
        codiceFiscale:
          type: ["string", "null"]
        indirizzo:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
    ClienteStorico:
      type: object
      properties:
        cliente:
          $ref: "#/components/schemas/Cliente"
        prenotazioni:
          type: array
          items:
            $ref: "#/components/schemas/Prenotazione"
        contratti:
          type: array
          items:
            $ref: "#/components/schemas/Contratto"
      required:
        - cliente
        - prenotazioni
        - contratti
    Prenotazione:
      type: object
      properties:
        id:
          type: integer
        vetturaId:
          type: integer
        clienteId:
          type: integer
        dataInizio:
          type: string
          format: date
        dataFine:
          type: string
          format: date
        stato:
          type: string
          description: "attiva | completata | annullata | in_corso"
        note:
          type: ["string", "null"]
        vettura:
          $ref: "#/components/schemas/Vettura"
        cliente:
          $ref: "#/components/schemas/Cliente"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - vetturaId
        - clienteId
        - dataInizio
        - dataFine
        - stato
        - vettura
        - cliente
        - createdAt
        - updatedAt
...
    Contratto:
      type: object
      properties:
        id:
          type: integer
        vetturaId:
          type: integer
        clienteId:
          type: integer
        prenotazioneId:
          type: ["integer", "null"]
        numero:
          type: string
        tipo:
          type: string
          description: "vendita | noleggio | leasing | permuta"
        dataContratto:
          type: string
          format: date
        importo:
          type: ["number", "null"]
        archiviato:
          type: boolean
        note:
          type: ["string", "null"]
        vettura:
          $ref: "#/components/schemas/Vettura"
        cliente:
          $ref: "#/components/schemas/Cliente"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - vetturaId
        - clienteId
        - numero
        - tipo
        - dataContratto
        - archiviato
        - vettura
        - cliente
        - createdAt
        - updatedAt
    CreateContrattoBody:
      type: object
      properties:
        vetturaId:
          type: integer
        clienteId:
          type: integer
        prenotazioneId:
          type: ["integer", "null"]
        numero:
          type: string
        tipo:
          type: string
        dataContratto:
          type: string
          format: date
        importo:
          type: ["number", "null"]
        archiviato:
          type: boolean
        note:
          type: ["string", "null"]
      required:
        - vetturaId
        - clienteId
        - numero
        - tipo
        - dataContratto
        - archiviato
    UpdateContrattoBody:
      type: object
      properties:
        numero:
          type: ["string", "null"]
        tipo:
          type: ["string", "null"]
        dataContratto:
          type: ["string", "null"]
        importo:
          type: ["number", "null"]
        archiviato:
          type: ["boolean", "null"]
        note:
          type: ["string", "null"]
...