# lib/api-zod/src/generated/types/clienteStorico.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Gestione Prenotazioni Concessionaria
 * OpenAPI spec version: 0.1.0
 */
import type { Cliente } from "./cliente";
import type { Contratto } from "./contratto";
import type { Prenotazione } from "./prenotazione";

export interface ClienteStorico {
  cliente: Cliente;
  prenotazioni: Prenotazione[];
  contratti: Contratto[];
}