# lib/api-zod/src/generated/types/contratto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Gestione Prenotazioni Concessionaria
 * OpenAPI spec version: 0.1.0
 */
import type { Cliente } from "./cliente";
import type { Vettura } from "./vettura";

export interface Contratto {
  id: number;
  vetturaId: number;
  clienteId: number;
  /** @nullable */
  prenotazioneId?: number | null;
  numero: string;
  /** vendita | noleggio | leasing | permuta */
  tipo: string;
  dataContratto: Date;
  /** @nullable */
  importo?: number | null;
  archiviato: boolean;
  /** @nullable */
  note?: string | null;
  vettura: Vettura;
  cliente: Cliente;
  createdAt: Date;
  updatedAt: Date;
}