# lib/api-zod/src/generated/types/createClienteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Gestione Prenotazioni Concessionaria
 * OpenAPI spec version: 0.1.0
 */

export interface CreateClienteBody {
  nome: string;
  cognome: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  telefono?: string | null;
  /** @nullable */
  codiceFiscale?: string | null;
  /** @nullable */
  indirizzo?: string | null;
  /** @nullable */
  note?: string | null;
}