# lib/api-zod/src/generated/types/createContrattoBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Gestione Prenotazioni Concessionaria
 * OpenAPI spec version: 0.1.0
 */

export interface CreateContrattoBody {
  vetturaId: number;
  clienteId: number;
  /** @nullable */
  prenotazioneId?: number | null;
  numero: string;
  tipo: string;
  dataContratto: Date;
  /** @nullable */
  importo?: number | null;
  archiviato: boolean;
  /** @nullable */
  note?: string | null;
}