# lib/api-zod/src/generated/types/updateClienteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Gestione Prenotazioni Concessionaria
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateClienteBody {
  /** @nullable */
  nome?: string | null;
  /** @nullable */
  cognome?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  telefono?: string | null;
  /** @nullable */
  codiceFiscale?: string | null;
  /** @nullable */
  indirizzo?: string | null;
  /** @nullable */
  note?: string | null;
}