# lib/api-zod/src/generated/types/updateContrattoBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Gestione Prenotazioni Concessionaria
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateContrattoBody {
  /** @nullable */
  numero?: string | null;
  /** @nullable */
  tipo?: string | null;
  /** @nullable */
  dataContratto?: string | null;
  /** @nullable */
  importo?: number | null;
  /** @nullable */
  archiviato?: boolean | null;
  /** @nullable */
  note?: string | null;
}